deploy/pyproject.toml
[project]
name = "consultoria"
version = "0.1.0"
description = "Agente PostgreSQL para consultas estruturadas em ambiente hospitalar"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.45.0",
    "anthropic>=0.50.0",
    "python-dotenv>=1.1.0",
    "agno>=2.5.10",
    "sqlalchemy>=2.0.48",
    "psycopg[binary]>=3.2.3",
    "openai>=2.29.0",
    "httpx>=0.28.1",
    "requests>=2.32.0",
    "uvicorn>=0.42.0",
    "fastapi>=0.135.2",
]
 
